// client/types/ApiClient.ts
type AvailableRadarKey = 2 | 3 | 6 | 7 | 10 | 11;
type RadarData = Record<AvailableRadarKey, number>;
type BonanzaTheme =
  | "Trading"
  | "Airdrop"
  | "Mining"
  | "Influencing"
  | "Connecting"
  | "BuidLing";
type BonanzaRoleLabel =
  | "WenLambo"
  | "AirdropBeacon"
  | "GoldenMiner"
  | "Superstar"
  | "HumanRouter"
  | "WinterWarrior";

type BonanzaRole = {
  key: AvailableRadarKey;
  type: BonanzaTheme;
  label: BonanzaRoleLabel;
};
// ...
type LuckyMonthValues = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;

type LuckyMonthsResponse = {
  maxMonth: string;
  monthScores: Record<LuckyMonthValues, number>;
  sumMonthScore: number;
};

type FortuneThemes =
  | "1-1"
  | "1-2"
  | "1-3"
  | "1-4"
  | "1-5"
  | "2-1"
  | "2-2"
  | "2-3"
  | "2-4"
  | "2-5"
  | "2-6"
  | "2-7"
  | "2-8"
  | "2-9"
  | "2-10"
  | "2-11"
  | "2-12";